+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2025-11-19 11:01:38
+
+### Changed Files
+- `src/features/dashboard/DashboardHome.tsx`
+
+### Statistics
+```
+ 1 file changed, 6 insertions(+), 17 deletions(-)
+```
+
 ## [Checkpoint] - 2025-11-19 10:27:42
 
 ### Changed Files